[refactor] refactoring
Model/Pieces -> Domain/Pieces/Piece
