+++ ChessAPIServer/Domain/Pieces/Knight.cs
@@ -1,0 +1,10 @@
+namespace APIServer.Domain.Pieces;
+public class Knight : Piece
+{
+    public new const string code_name = "N";
+
+    public override bool MoveCheck(int index)
+    {
+        throw new NotImplementedException();
+    }
+}
