+++ ChessAPIServer/Domain/Pieces/Queen.cs
@@ -1,0 +1,10 @@
+namespace APIServer.Domain.Pieces;
+public class Queen : Piece
+{
+    public new const string code_name = "Q";
+
+    public override bool MoveCheck(int index)
+    {
+        throw new NotImplementedException();
+    }
+}
